# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitCheck",
    "description": "Resume/JD matching pipeline with explainable scoring.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Job",
        "description": "Queue a job-posting URL for processing.\n\nReturns **202 Accepted**, not 201. The distinction is the whole point of\nthis milestone: 201 would promise the work is done and the result is\navailable at the returned location. 202 promises only that the request\nwas accepted for processing -- which is exactly the truth, because\nfetching an arbitrary URL takes 200ms to 30s, fails often, and cannot\nhappen inside a request handler.\n\nSubmitting the same URL twice for the same profile returns the existing\njob rather than creating a duplicate scrape.",
        "operationId": "submit_job_api_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List jobs, newest first. Filterable by profile and status.\n\nBoth filters hit indexed columns. This is the query the M10 ops dashboard\nis built on -- `?status=dead` is the dead-letter list.\n\nThe ownership join is not optional and not a filter the caller supplies.\n`?profile_id=` narrows within what this user owns; it cannot widen beyond\nit, so passing someone else's id returns an empty list rather than their\njobs. A listing endpoint is the easiest place to leak everything at once,\nbecause one missing predicate returns every row in the table instead of\none.",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Current state of one job. This is the endpoint the frontend polls.\n\nSets `Cache-Control: no-store` because a cached job status is worse than\nuseless -- a proxy holding a `queued` response for 30 seconds makes the\nUI look stuck when the work has already finished.\n\nOwnership is inherited through the profile rather than stored on the job.\nA job belongs to whoever owns the resume it was submitted for, so joining\nkeeps one source of truth; a `user_id` column here could disagree with\n`profiles.user_id` and there would be no way to say which was right.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List Matches",
        "description": "Matches for one of the caller's profiles, best first.\n\nThis is the query `matches_feed_idx` exists for -- `(profile_id,\nfinal_score DESC)` serves both the filter and the ordering, so the\nplanner never sorts.\n\nOwnership is checked on the *profile* rather than filtered on the\nmatches. Filtering alone would return an empty list for someone else's\nprofile, which is a subtly worse answer: it says \"this profile has no\nmatches\" where the truth is \"this profile is not yours\", and a client\ncannot tell an empty feed from a forbidden one.\n\n**The filters here are not the ones in `retrieval.FeedFilters`, and the\nduplication is deliberate.** Those run at recall time and decide what gets\n*scored*; these run at read time and decide what gets *shown*. Collapsing\nthem would mean re-running a 200-candidate rerank every time somebody\nticked \"remote only\", which is a scoring job's worth of work on the\nrequest path -- exactly what section 6.9 puts on a queue instead.\n\n**Closed postings are hidden by default rather than deleted from the\nfeed.** A role filled after it was scored is still a true record of what\nwas recommended, and `include_closed` keeps it reachable; showing it\nunasked would be presenting a dead link as a live opportunity.",
        "operationId": "list_matches_api_matches_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "name": "remote_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Remote Only"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "max_min_years",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Min Years"
            }
          },
          {
            "name": "include_closed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Closed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResponse"
                  },
                  "title": "Response List Matches Api Matches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/saved": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Saved Matches",
        "description": "Everything the caller reacted to, most recent reaction first.\n\nDeclared *before* `/{match_id}`, and that ordering is load-bearing:\nFastAPI matches routes in declaration order, so the other way round this\npath would be captured by the id route and answered with a 422 about\n\"saved\" not being an integer.\n\n**One row per match, not one per verdict.** `match_feedback` is\nappend-only, so a posting marked interested and later applied has two\nrows; this is a tracker, and a tracker showing the same job twice at two\nstages is a worse tracker. `DISTINCT ON` takes the newest per match, which\nis the current state -- while the history stays intact underneath for the\nranking model that would eventually read it.",
        "operationId": "saved_matches_api_matches_saved_get",
        "parameters": [
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedMatch"
                  },
                  "title": "Response Saved Matches Api Matches Saved Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Match",
        "description": "One match with its full skill breakdown.\n\nJoined through `profiles` rather than checked afterwards, so a match\nbelonging to another user is indistinguishable from one that does not\nexist -- the row never leaves the database.",
        "operationId": "get_match_api_matches__match_id__get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/recommendations": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Build Recommendations",
        "description": "Ask for this profile's feed to be built, if it needs building.\n\nSection 6.9 option 1 -- lazy, with a warm state. The three properties that\nmake this the chosen strategy over a nightly sweep are all visible here:\nthe work is queued rather than run on the request path, it happens because\nsomebody asked rather than on a schedule, and the cost therefore tracks\nactive users rather than registered ones.\n\n**202 always, never 200 with results.** Building a feed is a recall plus\n200 reranks; doing it inline would make the first feed request the slowest\nrequest in the application and hold a worker-equivalent of work open on a\nweb connection. The client gets an acknowledgement and polls the feed it\nwas already polling.\n\n**`already_current` is the interesting return.** A feed that was scored\nunder the current `scorer_version` does not need rebuilding, so a client\nthat polls this endpoint cannot stampede the queue -- which matters\nbecause the natural client implementation calls it whenever the feed looks\nempty, and an empty feed is exactly the state a user stares at.",
        "operationId": "build_recommendations_api_matches_recommendations_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "profile_id": {
            "type": "integer",
            "title": "Profile Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "dead"
            ],
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_terminal": {
            "type": "boolean",
            "title": "Is Terminal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "url",
          "status",
          "attempts",
          "created_at",
          "updated_at",
          "is_terminal"
        ],
        "title": "JobResponse",
        "description": "A job's current state. This is what the frontend polls."
      },
      "JobSubmitRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "profile_id": {
            "type": "integer",
            "title": "Profile Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "url",
          "profile_id"
        ],
        "title": "JobSubmitRequest",
        "description": "Body of POST /api/jobs.\n\n`HttpUrl` is doing real work here: a malformed URL is rejected with a\nfield-level 422 before any row is written or any job enqueued. Without\nit, the bad value reaches a worker minutes later and fails there, which\nis a far worse place to discover a typo."
      },
...
    }
  }
}